requirements.txt
PyQt5>=5.15.0
wmi>=1.5.1
pywin32>=228
psutil>=5.8.0
pyserial>=3.5
pandas>=1.3.0
numpy>=1.19.0
matplotlib>=3.4.0
pytest>=6.2.0
pyyaml>=5.4.0 